requirements.txt
bs4
pyfiglet
pyTelegramBotAPI
python-dotenv
requests
